+++ src/stores/chains.ts
@@ -7,6 +7,7 @@ function state() {
     id: "", // chainId, this name is replaced when addign a chain
     liveEventIds: [] as string[], // the chain of event nodettes , ["showSpeech_abcd", "wait_abcd"]. when it's empty, the chain'sRemoved
     // addTime?
+    canAutoActivate: true, // whether events in the chain can be auto started, starts as false for addSubEvents, and can be set to true to activate
   };
 }